requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
currencies==2020.12.12
Django==3.2
django-countries==7.3.2
django-currencies==0.10.1
django-debug-toolbar==3.4.0
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
importlib-metadata==4.11.4
itypes==1.2.0
Jinja2==3.1.2
jsonfield==2.1.1
Markdown==3.3.7
MarkupSafe==2.1.1
openapi-codec==1.3.2
phonenumbers==8.12.50
Pillow==9.1.1
PyJWT==1.7.1
pytz==2022.1
pyuca==1.2
requests==2.27.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
